反射，切面日志

+++ src/main/java/com/followme/model/User.java
@@ -6,9 +6,9 @@ package com.followme.model;
 public class User {
     private String userName;
     private Integer sex;
-    private Integer age;
+    private Long age;
 
-    public User(String userName, Integer sex, Integer age) {
+    public User(String userName, Integer sex, Long age) {
         this.userName = userName;
         this.sex = sex;
         this.age = age;
@@ -30,11 +30,11 @@ public class User {
         this.sex = sex;
     }
 
-    public Integer getAge() {
+    public Long getAge() {
         return age;
     }
 
-    public void setAge(Integer age) {
+    public void setAge(Long age) {
         this.age = age;
     }
 }
